// plio-api/burdi/apis/TopicController.ts
/* eslint-disable */

/**
 * Burdi.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { createEndpoint, normalizePath } from '@1023-ventures/darri-core';

import {
    ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';
import {
    Topic,
    TopicFromJSON,
    TopicToJSON,
} from '../models/Topic';

// Data Transfer Objects for individual endpoints

type DeleteUrlParams = {
    /** id  true*/
    id: string;
}

type DeleteQueryParams = {
    /**  */
    concurrency?: string;
}

type CreateBody =
    /**  */
    & Topic;

type CreateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

type ReadUrlParams = {
    /** id  true*/
    id: string;
}

type SparseUpdateBody =
    /**  */
    & string;

type SparseUpdateUrlParams = {
    /** id  true*/
    id: string;
    /** reload  true*/
    reload: boolean;
}

type UpdateBody =
    /**  */
    & Topic;

type UpdateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

/**
 * no description
 */
export const TopicController = {

    /**
     */
    _delete: createEndpoint<boolean, never, DeleteUrlParams, DeleteQueryParams>({
        method: 'DELETE',
        url: normalizePath('/burdi/api/Topic/{id}')
    }),
    /**
     */
    create: createEndpoint<Topic, CreateBody, CreateUrlParams, never>({
        method: 'POST',
        url: normalizePath('/burdi/api/Topic')
    }),
    /**
     */
    read: createEndpoint<Topic, never, ReadUrlParams, never>({
        method: 'GET',
        url: normalizePath('/burdi/api/Topic/{id}')
    }),
    /**
     */
    readAll: createEndpoint<Array<Topic>, never, never, never>({
        method: 'GET',
        url: normalizePath('/burdi/api/Topic')
    }),
    /**
     */
    sparseUpdate: createEndpoint<Topic, SparseUpdateBody, SparseUpdateUrlParams, never>({
        method: 'PATCH',
        url: normalizePath('/burdi/api/Topic/{id}')
    }),
    /**
     */
    update: createEndpoint<Topic, UpdateBody, UpdateUrlParams, never>({
        method: 'PUT',
        url: normalizePath('/burdi/api/Topic')
    }),
}